refactor: updateDragState 함수 util/drag.ts 이동
